Re-instate support for ios simulator tests (#7)
Re-instate support for running tests on ios simulator, now that we have separate ios simulator builds.
Re-instate the ios_simulator_launcher.py script but adapt for changes that have occurred: no longer
use ios-sim tool as that no longer works but replace with a shell script that uses xcrun; and reflect
fact the the gtest output has changed a bit.